// mobile/src/client/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type Project = {
    /**
     * The unique identifier of the project
     */
    id: string;
    /**
     * The name of the project
     */
    name: string;
    /**
     * The description of the project
     */
    description?: string;
    /**
     * The current status of the project
     */
    status: 'Backlog' | 'To Do' | 'In Progress' | 'Completed';
    assignee?: User;
    /**
     * The date and time when the project was created
     */
    createdAt: string;
    /**
     * The date and time when the project was last updated
     */
    updatedAt: string;
};

export type User = {
    /**
     * The unique identifier of the user
     */
    id: string;
    /**
     * The name of the user
     */
    name: string;
    /**
     * The email address of the user
     */
    email: string;
    /**
     * The URL of the user's avatar image
     */
    avatar?: string;
    /**
     * The date and time when the user was created
     */
    createdAt: string;
    /**
     * The date and time when the user was last updated
     */
    updatedAt: string;
};

export type GetApiProjectsData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/api/projects';
};

export type GetApiProjectsResponses = {
    /**
     * List of projects
     */
    200: Array<Project>;
};

export type GetApiProjectsResponse = GetApiProjectsResponses[keyof GetApiProjectsResponses];

export type PostApiProjectsData = {
    body: {
        name: string;
        description: string;
        status: 'Backlog' | 'To Do' | 'In Progress' | 'Completed';
        /**
         * The ID of the user to assign the project to
         */
        assignee_id?: string;
    };
    path?: never;
    query?: never;
    url: '/api/projects';
};

export type PostApiProjectsResponses = {
    /**
     * The created project
     */
    201: Project;
};

export type PostApiProjectsResponse = PostApiProjectsResponses[keyof PostApiProjectsResponses];

export type PutApiProjectsByIdData = {
    body: {
        name: string;
        description: string;
        status: 'Backlog' | 'To Do' | 'In Progress' | 'Completed';
        /**
         * The ID of the user to assign the project to
         */
        assignee_id: string;
    };
    path: {
        /**
         * The ID of the project to update
         */
        id: unknown;
    };
    query?: never;
    url: '/api/projects/{id}';
};

export type PutApiProjectsByIdResponses = {
    /**
     * The updated project
     */
    200: Project;
};

export type PutApiProjectsByIdResponse = PutApiProjectsByIdResponses[keyof PutApiProjectsByIdResponses];

export type GetApiUsersData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/api/users';
};

export type GetApiUsersResponses = {
    /**
     * List of users
     */
    200: Array<User>;
};

export type GetApiUsersResponse = GetApiUsersResponses[keyof GetApiUsersResponses];

export type ClientOptions = {
    baseURL: 'http://localhost:3000' | (string & {});
};